added conditions and closing database

+++ app/infrastructure/repositories/order_mysql_repository.py
@@ -17,14 +17,20 @@ except Exception as e:
     
 class OrderMysqlRepository(IOrderRepository):
     def read_orders(self):
-        mycursor.execute("SELECT * FROM orders")
-        orders = mycursor.fetchall()
-        order_list=[]
-        for order in orders:
-            new_order = Order(id=order[0], user_id=order[1], street_name=order[2], postal_code=order[3], city=order[4], 
-                              date_of_order=order[5], total_price=order[6], tax=order[7])
-            order_list.append(new_order)
-        return order_list
+        try:
+            mycursor.execute("SELECT * FROM orders")
+            orders = mycursor.fetchall()
+            order_list=[]
+            for order in orders:
+                new_order = Order(id=order[0], user_id=order[1], street_name=order[2], postal_code=order[3], city=order[4], 
+                                date_of_order=order[5], total_price=order[6], tax=order[7])
+                order_list.append(new_order)
+            return order_list
+        finally:
+            if mycursor:
+                mycursor.close()
+            if mydb.is_connected():
+                mydb.close()
     
     def read_order(self, order_id: int):
         try:
@@ -46,22 +52,28 @@ class OrderMysqlRepository(IOrderRepository):
                 mydb.close()
     
     def create_order(self, new_values: OrderValues):
-        sql = ("INSERT INTO orders(user_id, street_name, postal_code, city, date_of_order, total_price, tax)"
-                "values(%s, %s, %s, %s, %s, %s, %s)")
-        val = (new_values.user_id, new_values.street_name, new_values.postal_code, new_values.city,
-               new_values.date_of_order, new_values.total_price, new_values.tax)
-        mycursor.execute(sql, val)
-        mydb.commit()
-        
-        mycursor.execute("SELECT LAST_INSERT_ID();")
-        last_id = mycursor.fetchone()
-        if last_id is not None and isinstance(last_id[0], int):
-            last_id_int : int = last_id[0]
-        mycursor.execute("SELECT * FROM orders WHERE ID = %s", (last_id_int,))
-        order = mycursor.fetchone()
-        new_order = Order(id=order[0], user_id=order[1], street_name=order[2], postal_code=order[3], city=order[4],
-                          date_of_order=order[5], total_price=order[6], tax=order[7]) if order is not None else None
-        return new_order
+        try:
+            sql = ("INSERT INTO orders(user_id, street_name, postal_code, city, date_of_order, total_price, tax)"
+                    "values(%s, %s, %s, %s, %s, %s, %s)")
+            val = (new_values.user_id, new_values.street_name, new_values.postal_code, new_values.city,
+                new_values.date_of_order, new_values.total_price, new_values.tax)
+            mycursor.execute(sql, val)
+            mydb.commit()
+            
+            mycursor.execute("SELECT LAST_INSERT_ID();")
+            last_id = mycursor.fetchone()
+            if last_id is not None and isinstance(last_id[0], int):
+                last_id_int : int = last_id[0]
+            mycursor.execute("SELECT * FROM orders WHERE ID = %s", (last_id_int,))
+            order = mycursor.fetchone()
+            new_order = Order(id=order[0], user_id=order[1], street_name=order[2], postal_code=order[3], city=order[4],
+                            date_of_order=order[5], total_price=order[6], tax=order[7]) if order is not None else None
+            return new_order
+        finally:
+            if mycursor:
+                mycursor.close()
+            if mydb.is_connected():
+                mydb.close()
     
     def delete_order(self, order_id: int):
         try:
@@ -75,12 +87,29 @@ class OrderMysqlRepository(IOrderRepository):
             id_exist = bool(id_exist)
             if not id_exist:
                 return None
-            sql = ("DELETE FROM orders WHERE ID = %s")
-            mycursor.execute(sql,val)
-            mydb.commit()
-            return id_exist
+            
+            sql2= ("SELECT EXISTS(SELECT 1 FROM orders_products WHERE order_id = %s)")
+            mycursor.execute(sql2, val)
+            result2 = mycursor.fetchone()
+            if result2 is None: #check if packed result is NONE, because unpacking NONE gives the error
+                return None
+            id_exist2, = result2 #unpacking the result, 
+            id_exist2 = bool(id_exist2)
+            if id_exist and id_exist2:
+                id_exist=409
+                return id_exist
+            elif id_exist and not id_exist2:
+                sql = ("DELETE FROM orders WHERE ID = %s")
+                mycursor.execute(sql,val)
+                mydb.commit()
+                return id_exist
         except:
             return None
+        finally:
+            if mycursor:
+                mycursor.close()
+            if mydb.is_connected():
+                mydb.close()
     
     def update_order(self, order_to_be_updated_id: int ,new_values: OrderValues):
         try:    
@@ -110,3 +139,8 @@ class OrderMysqlRepository(IOrderRepository):
             return new_order
         except:
             return None
+        finally:
+            if mycursor:
+                mycursor.close()
+            if mydb.is_connected():
+                mydb.close()
